+++ src/main/java/com/exaroton/api/UnsupportedProtocolException.java
@@ -1,7 +1,0 @@
-package com.exaroton.api;
-
-public class UnsupportedProtocolException extends Exception {
-    public UnsupportedProtocolException(String s) {
-        super(s);
-    }
-}