+++ src/test/java/ClientOptionsTest.java
@@ -1,8 +1,6 @@
-import com.exaroton.api.UnsupportedProtocolException;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class ClientOptionsTest extends APIClientTest {
     @Test
@@ -18,33 +16,6 @@ public class ClientOptionsTest extends APIClientTest {
     @Test
     void apiTokenNull() {
         assertThrows(IllegalArgumentException.class, () -> client.setAPIToken(null));
-    }
-
-    @Test
-    void protocolHTTP() {
-        assertDoesNotThrow(() -> client.setProtocol("http"));
-        assertDoesNotThrow(() -> client.setProtocol("HTTP"));
-    }
-
-    @Test
-    void protocolHTTPS() {
-        assertDoesNotThrow(() -> client.setProtocol("https"));
-        assertDoesNotThrow(() -> client.setProtocol("HTTPS"));
-    }
-
-    @Test
-    void protocolFTP() {
-        assertThrows(UnsupportedProtocolException.class, () -> client.setProtocol("ftp"));
-    }
-
-    @Test
-    void protocolEmpty() {
-        assertThrows(UnsupportedProtocolException.class, () -> client.setProtocol(""));
-    }
-
-    @Test
-    void protocolNull() {
-        assertThrows(IllegalArgumentException.class, () -> client.setProtocol(null));
     }
 
     @Test